+++ Hint.h
@@ -17,7 +17,7 @@ public:
 	void Render(float x, float y);
 
 	void Prepare(int state, int col1, int col2, int col3, int col4, Rect hintRect);
-	int Height() { return _mcText.Height(); }
+	int Height() { return static_cast<int>(_mcText.Height()); }
 
 protected:
 	int _hintIndex;
